+++ js/utils.js
@@ -1,0 +1,34 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const createIdFromRangeGenerator = () => {
+  let id = 0;
+
+  return function () {
+    id++;
+    return id;
+  };
+};
+
+export {
+  getRandomInteger,
+  createRandomIdFromRangeGenerator,
+  createIdFromRangeGenerator,
+};
